Make entry points for camera plugins
This solves the problem of where to look for the camera plugins. They
are now all described in the 'camera_plugins' entry point of the
'beams' package. Use pkg_resources to load the entry point and get the
plugin object.

Deleted all plugin info .py files, instead the plugin info is contained
in a 'plugin_info' member in the camera class itself. The ID and
module/class names are obsolete.

This has the added benefit that if a camera plugin's dependencies are
not installed, then that plugin is simply not available through the GUI.