Ability to select camera number in OpenCV plugin
When you select a different camera, the plugin should test if it can grab
a frame from the camera, and if not, revert back to the old camera
number.

This should allow switching between two connected webcams with the OpenCV
plugin, not just using the default webcam. (Often the default webcam on a
laptop is the built-in camera, not useful for beam profiling.)